+++ include/flamegpu/runtime/messaging/Array2D.h
@@ -214,7 +214,7 @@ class MsgArray2D {
              */
             inline __device__ iterator begin(void) const {
                 // Bin before initial bin, as the constructor calls increment operator
-                return iterator(*this, -radius, -radius-1);
+                return iterator(*this, -static_cast<int>(radius), -static_cast<int>(radius)-1);
             }
             /**
              * Returns an iterator to the position beyond the end of the message list subset
